some organization
added:
- Player Controller (place where all other controllers execute)

working on organization
