+++ Camera Game/Assets/Scripts/PlayerScripts.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a1804a1456fd7c44a8197f52665d57b2
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
